+++ packages/terra-cli/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 1.2.0 - (November 24, 2020)
+
+* Added
+  * Added additional logging context for when commands error out with some stderr information
+  
+* Changed
+  * Updated the doc to use a simpler import of the terra-cli logger
+
+## 1.1.0 - (November 19, 2020)
+
 * Added
   * Updated the logger to accept multiple messages