+++ packages/terra-toolkit-docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.2.0 - (November 24, 2020)
+
+* Updated
+  * Updated example involving usage of webpack merge.
+
 ## 1.1.0 - (October 20, 2020)
 
 * Changed